Refactoring & documentation ESP - Refactoreded wifiHandler - Added more documentation - Clean-up/Removed unused functions
