+++ Chas Advance ESP32/.gitignore
@@ -4,3 +4,4 @@
 .vscode/launch.json
 .vscode/ipch
 /include/ESPSECRETS.h
+/src/ESPSECRETS.cpp
