+++ Chas Advance ESP32/include/espLogger.h
@@ -3,7 +3,9 @@
 
 #include "FS.h"
 #include "LittleFS.h"
+#include <Arduino.h>
 #include <ArduinoJson.h>
+#include <vector>
 #include "sensorDataHandler.h"
 #include "log.h"
 #include <CRC32.h>
@@ -11,7 +13,12 @@
 #define MAX_ENTRIES 16
 
 /**
- * @brief Batch structure to store multiple sensor entries with CRC.
+ * @brief Persistent on-flash representation of a batch.
+ *
+ * Stored layout on disk:
+ *  - uint16_t numEntries
+ *  - SensorData[numEntries]
+ *  - uint32_t crc32 (CRC of entries array)
  */
 struct Batch
 {
@@ -20,123 +27,128 @@ struct Batch
     uint32_t crc32; /**< CRC of timestamp + numEntries + entries */
 };
 
+/**
+ * @brief Binary record used to log send attempts for batches.
+ *
+ * Stored compactly in /send_status.bin to save flash space.
+ */
 struct SendStatusEntry {
-    uint32_t timestamp;     // Unix time or millis()
-    int32_t batchId;        // Batch ID
-    bool success;           // true = OK, false = FAIL
-    char message[40];      // Fixed-length message (null-terminated)
+    uint32_t timestamp;     ///< Unix time or millis()
+    int32_t batchId;        ///< Batch ID
+    bool success;           ///< true = OK, false = FAIL
+    char message[40];       ///< Fixed-length message (null-terminated)
 };
 
+// ----------------------------------------------------------------------------
 /**
- * @brief ESPLogger handles error logging and batch sensor logging to LittleFS.
+ * @class ESPLogger
+ * @brief Helper for logging sensor batches and operational errors to LittleFS.
+ *
+ * Responsibilities:
+ *  - Persist incoming sensor batches as binary files (/batch_<idx>.bin).
+ *  - Maintain CRC for batch integrity and validate on read.
+ *  - Record send-status events in a compact binary log.
+ *  - Provide simple human-readable dumps to Serial for debugging.
+ *  - Manage storage limits to MAX_BATCHES.
  */
 class ESPLogger
 {
 public:
-    /** @brief Constructor */
+    /** @brief Constructor. */
     ESPLogger();
 
     /**
-     * @brief Initializes LittleFS filesystem and prints status to Serial.  
+     * @brief Initialize LittleFS. Must be called once before other operations.
+     *
+     * Prints mount status to Serial.
      */
     void begin();
 
     // -------- Error logging --------
     /**
-     * @brief Logs an error message to the error file with timestamp.
-     * @param msg Error message string.
+     * @brief Append an error message to the error log with a timestamp.
+     * @param msg Human readable error message
      */
     void logError(const String &msg);
 
-    /**
-     * @brief Prints all saved error messages to Serial.
-     */
+    /** @brief Print all saved error messages to Serial (if any). */
     void printErrors();
 
-    /**
-     * @brief Clears all stored error messages.
-     */
+    /** @brief Remove the error log file. */
     void clearErrors();
 
     // -------- Batch logging --------
     /**
-     * @brief Logs a batch of sensor readings to LittleFS.
-     * @param arr JSON array of sensor objects (must contain temperature, humidity, error, errorType)
+     * @brief Store a vector of SensorData as a new batch file.
+     *
+     * The function writes: count (uint16_t) + raw entries + crc32(uint32_t).
+     *
+     * @param entries Vector of SensorData to persist. Must not be empty.
      */
-    void logBatch(std::vector<SensorData>& entries);
+    void logBatch(std::vector<SensorData> &entries);
 
-    /**
-     * @brief Prints all valid batch files and their entries to Serial.
-     */
+    /** @brief Print all valid batches and their entries to Serial. */
     void printBatches();
 
     /**
-     * @brief Retrieves the oldest valid batch from storage.
-     * @param outEntries Vector to store sensor entries.
-     * @param batchIndex Index of the batch file.
-     * @return True if a valid batch was retrieved, false otherwise.
+     * @brief Retrieve the oldest valid batch from storage.
+     * @param outEntries Output vector filled with the batch entries.
+     * @param batchIndex Output index of the returned batch file.
+     * @return true when a valid batch was found and populated.
      */
     bool getOldestBatch(std::vector<SensorData> &outEntries, uint16_t &batchIndex);
 
     /**
-     * @brief Retrieves the newest valid batch from storage.
-     * @param outEntries Vector to store sensor entries.
-     * @param batchIndex Index of the batch file.
-     * @return True if a valid batch was retrieved, false otherwise.
+     * @brief Retrieve the newest valid batch from storage.
+     * @param outEntries Output vector filled with the batch entries.
+     * @param batchIndex Output index of the returned batch file.
+     * @return true when a valid batch was found and populated.
      */
     bool getNewestBatch(std::vector<SensorData> &outEntries, uint16_t &batchIndex);
 
-    /**
-     * @brief Removes the oldest batch file from storage.
-     */
+    /** @brief Delete the oldest batch file from storage. */
     void removeOldestBatch();
 
-    /**
-     * @brief Deletes all batch files from storage.
-     */
+    /** @brief Remove all batch files. */
     void clearBatches();
 
-    /**
-     * @brief Log send status of a batch 
-     */
+    /** @brief Log the result of an attempt to send a batch to the backend. */
     void logSendStatus(int batchId, bool success, const String &message);
 
-    /**
-     * @brief Print send status logs to Serial
-     */
+    /** @brief Print the send-status log to Serial (human readable). */
     void printSendStatusLogs();
 
     // -------- Batch file utilities --------
     /**
-     * @brief Generates the filename for a batch index.
-     * @param index Batch index number.
-     * @return Full path string for the batch file.
+     * @brief Get batch filename for a given index.
+     * @param index Batch index
+     * @return Filename string (e.g. "/batch_3.bin")
      */
     String getBatchFilename(uint16_t index);
 
     /**
-     * @brief Returns a sorted list of all batch indices currently stored.
-     * @return Vector of batch indices.
+     * @brief Collect and sort stored batch indices.
+     * @return Sorted vector of batch indices currently stored.
      */
     std::vector<uint16_t> getBatchIndices();
 
     /**
-     * @brief Reads a batch file, validates CRC and returns sensor entries.
-     * @param fname Filename of the batch.
-     * @param outEntries Vector to store sensor entries.
-     * @return True if the batch is valid and read successfully.
+     * @brief Read a batch file and validate CRC before populating outEntries.
+     * @param fname Filename to read
+     * @param outEntries Output vector populated on success
+     * @return true if the file was present, read and CRC matched
      */
     bool readBatchFile(const String &fname, std::vector<SensorData> &outEntries);
 
     /**
-     * @brief Prints a vector of sensor entries to Serial.
-     * @param entries Vector of SensorEntry objects.
+     * @brief Pretty-print an in-memory vector of SensorData to Serial.
+     * @param entries Vector of SensorData
      */
     void printEntries(const std::vector<SensorData> &entries);
 
 private:
     const char *ERROR_FILE = "/errors.txt";  /**< Path to error log file */
-    const size_t MAX_BATCHES = 20;           /**< Maximum number of batch files stored */
+    const size_t MAX_BATCHES = 20;           /**< Max number of batch files kept */
 };
 
-#endif
+#endif // ESPLOGGER_H
