+++ Chas Advance ESP32/include/sensorDataHandler.h
@@ -5,18 +5,22 @@
 #include <ArduinoJson.h>
 #include <vector>
 
+/**
+ * @brief Packed binary representation of a single sensor reading.
+ *
+ * The struct is packed to ensure stable on-wire / on-flash layout when
+ * writing/reading binary batches.
+ */
 #pragma pack(push, 1)
 struct SensorData
 {
-    uint8_t SensorId;    /**< ID of the sensor */
-    uint32_t timestamp;   /**< Unix timestamp of the sensor reading */
-    float temperature;    /**< Temperature value */
-    float humidity;       /**< Humidity value */
-    bool error;           /**< Whether there was an error in this reading */
-    uint8_t errorType;    /**< Type of error, if any */
+    uint8_t SensorId;    /**< ID of the sensor (0-255) */
+    uint32_t timestamp;  /**< Unix timestamp (seconds) or sensor-local millis depending on context */
+    float temperature;   /**< Temperature in degrees Celsius */
+    float humidity;      /**< Relative humidity in percent */
+    bool error;          /**< True if this reading indicates an error */
+    uint8_t errorType;   /**< Error type code (see ErrorType enum in log.h) */
 };
 #pragma pack(pop)
-float median(std::vector<float>& values);
-SensorData calcMedian(JsonArray& arr);
 
 #endif
