+++ Chas Advance ESP32/include/wifiHandler.h
@@ -1,7 +1,23 @@
+/**
+ * @file wifiHandler.h
+ * @brief HTTP ingestion, WiFi and batch-handling public API for the ESP32 gateway.
+ *
+ * This header exposes functions used to initialize and run the network-facing
+ * parts of the firmware: STA/AP setup, a simple HTTP server that accepts POST
+ * /data binary payloads and queue-based batch processing + retry helpers.
+ *
+ * The implementation is designed to:
+ *  - accept binary batches from sensors (sendMillis + SensorData[]),
+ *  - enqueue them for processing by a background task,
+ *  - persist failed batches and retry sending to backend.
+ *
+ * Note: many helper functions operate on Arduino/ESP32-specific types
+ * (WiFiClient, WiFiServer). This header is intended for firmware builds.
+ */
+
 #ifndef WIFIHANDLER_H
 #define WIFIHANDLER_H
-#include <Arduino.h>
-#include <ArduinoJson.h>
+
 #include <WebServer.h>
 #include <NTPClient.h>
 #include <WiFiUdp.h>
@@ -9,32 +25,39 @@
 #include "log.h"
 #include <time.h>
 
-#define maxRequestBodySize (10*1024) // Max size of incoming POST body in bytes
-
-// initialize wifi setup
+/**
+ * @brief Initialize WiFi, access point and HTTP server, and configure NTP.
+ *
+ * Internally calls connectToWiFi(),
+ * setupAccessPoint(), setupHttpServer(), and configTime().
+ */
 void initWifi();
-// Connects the ESP32 to the WiFi network
+
+/**
+ * @brief Connect the ESP32 to configured STA WiFi network.
+ *
+ * Blocks until WiFi connected in current implementation.
+ */
 void connectToWiFi();
-// Sets up the Access Point for the Arduino to connect to
+
+/**
+ * @brief Configure the device as an Access Point for local device pairing.
+ *
+ * Starts a soft AP using ap_ssid/ap_password from secrets.
+ */
 void setupAccessPoint();
-// Sets up the HTTP server to handle incoming requests
+
+/**
+ * @brief Start HTTP server and attach routes.
+ *
+ * Currently opens server on port 80 and prepares for incoming POST /data.
+ */
 void setupHttpServer();
-// Handles incoming POST requests to /data
-void handleClientAsync();
-void processBatches(void *parameter);
-// POST request helpers
-bool isValidPostRequest(WiFiClient &client, const String &requestLine);
-int readContentLength(WiFiClient &client);
-bool readRequestBody(WiFiClient &client, std::vector<uint8_t> &buffer, int contentLength);
-bool parseBatch(const std::vector<uint8_t> &buffer, uint32_t &sendMillis, std::vector<SensorData> &batch);
-void respond(WiFiClient &client, int code); 
 
-// Retry sending saved batches
-void trySendPendingBatches();
-bool sendJsonToServer(const String &jsonString, int batchId);
-bool postBatchToServer(const std::vector<SensorData> &batch, int batchID);
+/**
+ * @brief Global server object that listens for incoming client connections.
+ * @note Defined in wifiHandler.cpp
+ */
+extern WiFiServer server; // Server listens to port 80
 
-extern unsigned long timeSinceDataReceived;
-extern WiFiServer server; // Server listen to port 80
-
-#endif
+#endif // WIFIHANDLER_H
