+++ Chas Advance ESP32/src/espLogger.cpp
@@ -1,7 +1,18 @@
 #include "espLogger.h"
 
+/**
+ * @brief Construct a new ESPLogger object.
+ *
+ * Lightweight constructor; real initialization happens in begin().
+ */
 ESPLogger::ESPLogger() {}
 
+/**
+ * @brief Initialize LittleFS and print mount status to Serial.
+ *
+ * Call this early in setup() before any file operations. On non-ESP32
+ * builds LittleFS may be unavailable — the header guards handle that.
+ */
 void ESPLogger::begin()
 {
     if (!LittleFS.begin(true))
@@ -15,6 +26,15 @@ void ESPLogger::begin()
 }
 
 // -------- Error logging --------
+
+/**
+ * @brief Append a timestamped error message to the persistent error log.
+ *
+ * The entry format is: "<timestamp> | <message>\n".
+ * If opening the error file fails, an error is printed to Serial.
+ *
+ * @param msg Human-readable error message to record.
+ */
 void ESPLogger::logError(const String &msg)
 {
     File f = LittleFS.open(ERROR_FILE, FILE_APPEND);
@@ -29,6 +49,11 @@ void ESPLogger::logError(const String &msg)
     f.close();
 }
 
+/**
+ * @brief Print the contents of the persistent error log to Serial.
+ *
+ * If no error log exists a short message is printed instead.
+ */
 void ESPLogger::printErrors()
 {
     File f = LittleFS.open(ERROR_FILE, FILE_READ);
@@ -46,13 +71,32 @@ void ESPLogger::printErrors()
     f.close();
 }
 
+/**
+ * @brief Remove the persistent error log file from storage.
+ *
+ * After calling this the error log will no longer be present until new
+ * errors are logged.
+ */
 void ESPLogger::clearErrors()
 {
     LittleFS.remove(ERROR_FILE);
 }
 
+/**
+ * @brief Persist a vector of SensorData as a new batch file on LittleFS.
+ *
+ * File layout:
+ *  - uint16_t count
+ *  - SensorData[count] (raw binary)
+ *  - uint32_t crc32 (CRC over the SensorData array)
+ *
+ * The function enforces MAX_BATCHES by removing the oldest file when needed.
+ * It logs failures to the error log.
+ *
+ * @param entries Vector of SensorData to persist. Must be non-empty.
+ */
 void ESPLogger::logBatch(std::vector<SensorData>& entries)
-   {
+{
     if (entries.empty())
     {
         Serial.println("No valid entries in JSON array; batch not logged");
@@ -97,6 +141,12 @@ void ESPLogger::logBatch(std::vector<SensorData>& entries)
                   nextIndex, count, (int)written);
 }
 
+/**
+ * @brief Print all stored, valid batches and their entries to Serial.
+ *
+ * Skips corrupted batch files (CRC mismatch or read errors) and prints a
+ * short notice for each skipped file.
+ */
 void ESPLogger::printBatches()
 {
     auto indices = getBatchIndices();
@@ -126,6 +176,16 @@ void ESPLogger::printBatches()
     Serial.println("---- Batch log end ----");
 }
 
+/**
+ * @brief Retrieve the oldest valid batch from storage.
+ *
+ * On success outEntries is populated and batchIndex is set to the file's index.
+ * If the oldest file is corrupted it will be removed and the function returns false.
+ *
+ * @param outEntries Output vector filled with the batch entries on success.
+ * @param batchIndex Output index of the returned batch file.
+ * @return true if a valid oldest batch was found; false otherwise.
+ */
 bool ESPLogger::getOldestBatch(std::vector<SensorData> &outEntries, uint16_t &batchIndex)
 {
     auto indices = getBatchIndices();
@@ -145,6 +205,15 @@ bool ESPLogger::getOldestBatch(std::vector<SensorData> &outEntries, uint16_t &ba
     return true;
 }
 
+/**
+ * @brief Retrieve the newest valid batch from storage.
+ *
+ * Similar to getOldestBatch but returns the most recent batch.
+ *
+ * @param outEntries Output vector filled with the batch entries on success.
+ * @param batchIndex Output index of the returned batch file.
+ * @return true if a valid newest batch was found; false otherwise.
+ */
 bool ESPLogger::getNewestBatch(std::vector<SensorData> &outEntries, uint16_t &batchIndex)
 {
     auto indices = getBatchIndices();
@@ -164,6 +233,11 @@ bool ESPLogger::getNewestBatch(std::vector<SensorData> &outEntries, uint16_t &ba
     return true;
 }
 
+/**
+ * @brief Remove the oldest batch file from storage.
+ *
+ * Useful to free space when the number of batch files reaches the configured limit.
+ */
 void ESPLogger::removeOldestBatch()
 {
     auto indices = getBatchIndices();
@@ -175,6 +249,11 @@ void ESPLogger::removeOldestBatch()
     LittleFS.remove(fname);
 }
 
+/**
+ * @brief Delete all stored batch files.
+ *
+ * Iterates over discovered batch indices and removes each file.
+ */
 void ESPLogger::clearBatches()
 {
     auto indices = getBatchIndices();
@@ -186,6 +265,16 @@ void ESPLogger::clearBatches()
     }
 }
 
+/**
+ * @brief Append a send-status record to the compact binary log.
+ *
+ * The log stores fixed-size SendStatusEntry structs. If the log grows beyond
+ * a configured maximum size it is cleared before appending.
+ *
+ * @param batchId Identifier of the batch associated with this send attempt.
+ * @param success true if the send succeeded; false otherwise.
+ * @param message Short human-readable message (will be truncated to fit).
+ */
 void ESPLogger::logSendStatus(int batchId, bool success, const String &message)
 {
     if (batchId < 0) return;
@@ -225,6 +314,11 @@ void ESPLogger::logSendStatus(int batchId, bool success, const String &message)
     sendStatusLogFile.close();
 }
 
+/**
+ * @brief Print the human-readable send-status log to Serial.
+ *
+ * Reads the binary log file and formats each SendStatusEntry into a readable line.
+ */
 void ESPLogger::printSendStatusLogs()
 {
     const char *filename = "/send_status.bin";
@@ -252,11 +346,28 @@ void ESPLogger::printSendStatusLogs()
 
 
 // -------- Batch file utilities --------
+
+/**
+ * @brief Construct the canonical filename for a batch index.
+ *
+ * Example: index=3 -> "/batch_3.bin"
+ *
+ * @param index Batch index number.
+ * @return String Filename for the batch file.
+ */
 String ESPLogger::getBatchFilename(uint16_t index)
 {
     return "/batch_" + String(index) + ".bin";
 }
 
+/**
+ * @brief Enumerate stored batch files and return their numeric indices.
+ *
+ * Scans the LittleFS root directory for filenames matching "/batch_<N>.bin",
+ * extracts the numeric part and returns a sorted list of indices.
+ *
+ * @return std::vector<uint16_t> Sorted indices of batch files.
+ */
 std::vector<uint16_t> ESPLogger::getBatchIndices()
 {
     std::vector<uint16_t> indices;
@@ -279,6 +390,17 @@ std::vector<uint16_t> ESPLogger::getBatchIndices()
     return indices;
 }
 
+/**
+ * @brief Read a batch file, validate its CRC and populate outEntries.
+ *
+ * Validates file size, reads the entry count, reads entries and stored CRC,
+ * computes CRC over read entries and compares. On any mismatch or read error
+ * the function returns false.
+ *
+ * @param fname Filename to read.
+ * @param outEntries Output vector populated with the entries on success.
+ * @return true if read and CRC validation succeeded; false otherwise.
+ */
 bool ESPLogger::readBatchFile(const String &fname, std::vector<SensorData> &outEntries)
 {
     File f = LittleFS.open(fname, FILE_READ);
@@ -329,6 +451,13 @@ bool ESPLogger::readBatchFile(const String &fname, std::vector<SensorData> &outE
     return true;
 }
 
+/**
+ * @brief Pretty-print an in-memory vector of SensorData to Serial.
+ *
+ * Each entry prints timestamp (formatted), temperature, humidity and error flags.
+ *
+ * @param entries Vector of SensorData to print.
+ */
 void ESPLogger::printEntries(const std::vector<SensorData> &entries)
 {
     for (const auto &entry : entries)
