+++ Chas Advance ESP32/src/log.cpp
@@ -1,7 +1,13 @@
 #include "log.h"
 
+/// Default NTP server hostname used by the system (definition for extern)
 const char *ntpServer = "pool.ntp.org";
 
+/**
+ * @brief Print a composed event line to Serial.
+ *
+ * Format: "<timestamp> <eventType> <description> <status>"
+ */
 void logEvent(String timestamp, String eventType, String description, String status)
 {
     Serial.print(timestamp);
@@ -13,6 +19,11 @@ void logEvent(String timestamp, String eventType, String description, String sta
     Serial.println(status);
 }
 
+/**
+ * @brief Format sensor values and emit via logEvent().
+ *
+ * Maps errorType to a suitable event level and message.
+ */
 void logSensorData(String timestamp, float temperature, float humidity, ErrorType errorType)
 {
     char buffer[50];
@@ -21,26 +32,45 @@ void logSensorData(String timestamp, float temperature, float humidity, ErrorTyp
     switch (errorType)
     {
     case NONE:
-        logEvent(timestamp,"INFO", buffer, "OK");
+        logEvent(timestamp, "INFO", buffer, "OK");
         break;
     case TOO_LOW:
-        logEvent(timestamp,"WARNING_Sensor data too low", buffer, "CHECK");
+        logEvent(timestamp, "WARNING_Sensor data too low", buffer, "CHECK");
         break;
     case TOO_HIGH:
-        logEvent(timestamp,"WARNING_Sensor data too high", buffer, "CHECK");
+        logEvent(timestamp, "WARNING_Sensor data too high", buffer, "CHECK");
         break;
     case SENSOR_FAIL:
-        logEvent(timestamp,"ERROR", "Sensor failure", "FAIL");
+        logEvent(timestamp, "ERROR", "Sensor failure", "FAIL");
+        break;
+    default:
+        logEvent(timestamp, "INFO", buffer, "UNKNOWN");
+        break;
     }
 }
 
-// OBS - Innehåller mockdata
+/**
+ * @brief Emit a startup/system reset message with current timestamp.
+ *
+ * Uses getTimeStamp() to obtain current local time. Intended to be called
+ * from setup() after time/NTP is configured.
+ */
 void logStartup()
 {
     String timeStamp = getTimeStamp();
     logEvent(timeStamp, "SYSTEM", "RESET", "OK");
 }
 
+/**
+ * @brief Check for prolonged absence of sensor data and log if needed.
+ *
+ * If the difference between current millis() and timeSinceDataReceived
+ * exceeds dataReceivedThreshold an error event is emitted and the
+ * timeSinceDataReceived is reset to the current millis() value to avoid
+ * repeated spamming of the same message.
+ *
+ * @param timeSinceDataReceived Reference to last-received millis value
+ */
 void checkDataTimeout(unsigned long &timeSinceDataReceived)
 {
     if ((millis() - timeSinceDataReceived) > dataReceivedThreshold)
@@ -51,6 +81,14 @@ void checkDataTimeout(unsigned long &timeSinceDataReceived)
     }
 }
 
+/**
+ * @brief Obtain a formatted timestamp string using NTP
+ *
+ * Attempts to read the current local time via getLocalTime(). On failure
+ * prints a diagnostic and returns "TIME_ERROR".
+ * 
+ * @return Timestamp in "YYYY-MM-DD HH:MM:SS" format or "TIME_ERROR".
+ */
 String getTimeStamp()
 {
     String timeStamp = "TIME_ERROR";
@@ -66,22 +104,37 @@ String getTimeStamp()
     {
         char buffer[20];
         strftime(buffer, sizeof(buffer), "%Y-%m-%d %H:%M:%S", &timeinfo);
-
         timeStamp = String(buffer);
     }
     return timeStamp;
 }
 
-//Convert unix timestamp to formatted string
+/**
+ * @brief Convert unix epoch seconds to a human-readable timestamp.
+ *
+ * Uses localtime() to produce a formatted string.
+ *
+ * @param ts Unix epoch seconds
+ * @return Formatted "YYYY-MM-DD HH:MM:SS" string
+ */
 String formatUnixTime(uint32_t ts)
 {
     time_t t = ts;
-    struct tm *timeinfo = localtime(&t); // UTC, no DST
+    struct tm *timeinfo = localtime(&t); // UTC
     char buffer[20];
     strftime(buffer, sizeof(buffer), "%Y-%m-%d %H:%M:%S", timeinfo);
     return String(buffer);
 }
 
+/**
+ * @brief Parse a human-readable timestamp into unix epoch seconds.
+ *
+ * Expected input format: "YYYY-MM-DD HH:MM:SS".
+ * On parse failure prints a diagnostic and returns 0.
+ *
+ * @param tsStr Timestamp string to parse
+ * @return Unix epoch seconds or 0 on failure
+ */
 uint32_t timestampStringToUnix(const String &tsStr)
 {
     struct tm timeinfo = {0};
@@ -100,6 +153,24 @@ uint32_t timestampStringToUnix(const String &tsStr)
     return (uint32_t)mktime(&timeinfo);
 }
 
+/**
+ * @brief Convert sensor-relative millis timestamps into absolute unix times.
+ *
+ * Sensors may deliver timestamps as their internal millis() values; this
+ * function uses the host's current unix time and the provided sendMillis
+ * (sensor millis when batch was sent) to compute absolute times
+ * for each entry. The function modifies batch[i].timestamp to hold unix
+ * epoch seconds for each measurement.
+ *
+ * Algorithm summary:
+ *  - Query current unix time (now) via getTimeStamp()
+ *  - Compute delay between last measurement's sensor-millis and sendMillis
+ *  - Compute unix time for latest measurement as now - delay_seconds
+ *  - Walk backwards converting prior entries using their millis deltas
+ *
+ * @param sendMillis 32-bit send time in milliseconds
+ * @param batch Vector of SensorData entries whose timestamp fields are updated
+ */
 void assignAbsoluteTimestamps(uint32_t sendMillis, std::vector<SensorData> &batch)
 {
   if (batch.empty())
@@ -112,15 +183,15 @@ void assignAbsoluteTimestamps(uint32_t sendMillis, std::vector<SensorData> &batc
   SensorData &latest = batch.back();
   uint32_t lastMeasurementMillis = latest.timestamp; // Arduino millis of last measurement
 
-  // How long since last measurement until batch was sent
+  // How long since last measurement until batch was sent (ms)
   uint32_t delayMs = (sendMillis >= lastMeasurementMillis) ? (sendMillis - lastMeasurementMillis) : 0;
 
   // Absolute Unix time of last measurement
   uint32_t lastMeasurementUnix = now - (delayMs / 1000);
   latest.timestamp = lastMeasurementUnix;
 
-  // Walk backwards for previous entries
-  for (int i = batch.size() - 2; i >= 0; i--)
+  // Walk backwards for previous entries and convert using measured deltas
+  for (int i = static_cast<int>(batch.size()) - 2; i >= 0; --i)
   {
     uint32_t deltaMs = lastMeasurementMillis - batch[i].timestamp; // millis between measurements
     batch[i].timestamp = lastMeasurementUnix - (deltaMs / 1000);
