+++ Chas Advance ESP32/src/mockJson.cpp
@@ -1,6 +1,7 @@
 #include "MockJson.h"
 #include <Arduino.h>
 
+//DEPRECATED
 String generateMockJson()
 {
   char buf[32];
